Keep previous commands the same
