+++ cmd/init.js
@@ -1,0 +1,6 @@
+
+function create(args) {
+
+}
+
+module.exports = create;
